Notification and Location services
Location service will show the user where they are relative to the nearest bubble tea shops that they can access to so that they can order at the nearest branch. Tracking of the personal location to be added.
Notification will receive messages from order microservice and will process it in order to send message to the customer.